Move steps to create symbol package into makefile

+++ sources/scripts/clean.ps1
@@ -32,7 +32,11 @@ Resolve-Path -Path "$SourcesRoot\Google.Solutions.*\obj" |
 	% { Remove-Item -Recurse -Force $_ }
 	
 # Delete obj directories
-Remove-Item -Recurse -Force "$SourcesRoot\installer\bin"
+if (Test-Path "$SourcesRoot\installer\bin") {
+	Remove-Item -Recurse -Force "$SourcesRoot\installer\bin"
+}
 
 # Delete packages
-Remove-Item -Recurse -Force "$SourcesRoot\packages"
+if (Test-Path "$SourcesRoot\packages") {
+	Remove-Item -Recurse -Force "$SourcesRoot\packages"
+}
